% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariatenorm.R
\name{bivariatenorm}
\alias{bivariatenorm}
\title{Bivariate Test of Normality}
\usage{
bivariatenorm(x1, x2)
}
\arguments{
\item{x1}{Univariate data variable}

\item{x2}{Univariate data variable}
}
\value{
Result of a bivariate t-test
}
\description{
A more concise version bivariatettest function that only contains output for a t-test. Function takes in 2 data variables, calculates Hotelling's T-square
and tests against 50% chi-square quantile. If 50% or more of the data lies within contour, then the pair of variables are normal.
}
\examples{
x1 <- rnorm(50)
x2 <- rnorm(50)
bivariatenorm(x1,x2)
}
