% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonf.andgir.cis.R
\name{bonf.andgir.cis}
\alias{bonf.andgir.cis}
\title{Bonferroni confidence interval for eigen values}
\usage{
bonf.andgir.cis(data, q, alpha)
}
\arguments{
\item{data}{A data set containing quantitative variables}

\item{q}{Number of eigen values we want to generate confidence intervals for}

\item{alpha}{Significance level}
}
\value{
List containing bonferonni confidence intervals
}
\description{
Function generates the bonferonni Anderson and Girshik confidence intervals for the eigen values of a data set
}
\examples{
df <- data.frame(x = rnorm(30), y = rnorm(30), z = rnorm(30))
bonf.andgir.cis(df,3,.10)
}
