% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorModelTest.R
\name{factorModelTest}
\alias{factorModelTest}
\title{Large Sample Test of Adequacy for an M-Factor Model}
\usage{
factorModelTest(x, alpha, ...)
}
\arguments{
\item{x}{An object of class factoAnalysis}

\item{alpha}{Significance level}

\item{...}{Extra parameters if needed}
}
\value{
A matrix with results of test
}
\description{
Function calculates the test statistic with Bartlett's correction against the chi-square quantile. If the test statistic is greater than the chi-square
quantile, then we reject the null that the M-factor model is appropriate for the data in favor of some other positive definite matrix. Results are then formatted into a gt table.
}
\examples{
\dontrun{
df = data.frame(x = rnorm(20), y = rnorm(20), z = rnorm(20))
obj <- pcfacta(df, m = 2),
factorModelTest(obj, .05)
}
}
