% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcabootAG.R
\name{pcabootAG}
\alias{pcabootAG}
\title{Bootstrapping of Principal Components}
\usage{
pcabootAG(data, alpha, iter, cov = TRUE)
}
\arguments{
\item{data}{A data set with quantitative variables}

\item{alpha}{Significance level}

\item{iter}{Number of bootstrap iterations}

\item{cov}{If TRUE uses S. Else, use R (correlation matrix).}
}
\value{
Histogram plots, boxplot, and output for bootstrapping of PCA and confidence intervals
}
\description{
This function performs bootstrapping on sample principal components of a given data set with the package "boot" and creates density plots for the eigenvalues and eigenvectors.
}
\examples{
df <- data.frame(x = rnorm(50), y = rnorm(50))
pcabootAG(df,.05,10000)
}
