% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaCalc.R
\name{pcaCalc}
\alias{pcaCalc}
\title{Principal Component Analysis Calculation}
\usage{
pcaCalc(data)
}
\arguments{
\item{data}{A data set containing quantitative variables}
}
\value{
A list containing sample covariance matrix, sample principal components, total sample variance, and a correlation matrix
}
\description{
This function performs principal component analysis. It constructs a sample covariance matrix, finds sample principal components, total sample variance,
and a cor(yi,xk) matrix.
}
\examples{
df <- data.frame(x = rnorm(30), y = rnorm(30), z = rnorm(30))
pcaCalc(df)
}
