% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcfacta.R
\name{pcfacta}
\alias{pcfacta}
\title{Principal Component Factor Analysis}
\usage{
pcfacta(data, m, R = TRUE)
}
\arguments{
\item{data}{A data frame}

\item{m}{The number of factors in the model}

\item{R}{Set to default TRUE meaning correlation matrix is used. If false, covariance matrix is used}
}
\value{
A list of class factoAnalysis containing loadings, communalities, residual matrix, specific variance
}
\description{
Function performs factor analysis. User can choose correlation or covariance matrix. Loadings, communalities, residual matrix, and tables are returned. This function also serves
as a constructor for analysis pertaining to FA.
}
\examples{
\dontrun{
df <- data.frame(x = rnorm(30), y = rnorm(30))
MATH5793BANH::pcfacta(df, m = 1)
}
}
