% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.factoAnalysis.R
\name{plot.factoAnalysis}
\alias{plot.factoAnalysis}
\title{Factor Analysis Plots}
\usage{
\method{plot}{factoAnalysis}(x, ...)
}
\arguments{
\item{x}{An object of factoAnalysis class}

\item{...}{Extra parameters if needed}
}
\value{
Plots of communality, specific variance, and pairs plots. List of loadings from varimax and promax rotations as well as reactable table.
}
\description{
Function overrides the base plot function for objects of factoAnalysis class. Function creates communality, specific variance, tables, and pairs plots of all the factors.
}
\examples{
\dontrun{
df <- data.frame(x = rnorm(30), y = rnorm(30))
obj <- pcfacta(df, m = 1)
plot(obj)
}
}
