% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proptest.R
\name{proptest}
\alias{proptest}
\title{Univariate Proportion Test of Normality}
\usage{
proptest(x)
}
\arguments{
\item{x}{Univariate data}
}
\value{
Proportions with 1/2 standard deviations of data set along with result of test.
}
\description{
This function takes in a data vector/univariate data set and calculates the proportion of data within 1/2 standard deviations of
the mean. Then, if the proportion falls within an equality then there are outliers present.
}
\examples{
x <- c(34,14,151,131,50,76)
proptest(x)

}
