% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcellipse.R
\name{qcellipse}
\alias{qcellipse}
\title{Quality Control Ellipse for Bivariate Data}
\usage{
qcellipse(x1, x2, alpha = 0.05)
}
\arguments{
\item{x1}{First data variable}

\item{x2}{Second data variable}

\item{alpha}{Rejection level}
}
\value{
A ggplot of a confidence ellipse
}
\description{
Function takes in 2 variables in which it calculates the (1-alpha)% quality control ellipse using ggplot's built-in function stat_ellipse.
}
\examples{
x1 <- rnorm(20)
x2 <- rnorm(20)
qcellipse(x1,x2)
}
