% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqcorr.R
\name{qqcorr}
\alias{qqcorr}
\title{Correlation Coefficient for a Q-Q plot}
\usage{
qqcorr(x)
}
\arguments{
\item{x}{Univariate data vector}
}
\value{
A correlation coefficient value -- cannot exceed 1
}
\description{
Correlation Coefficient for a Q-Q plot
}
\examples{
x <- rnorm(20)
qqcorr(x)
}
