% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyfactoranalysis.R
\name{shinyfactoranalysis}
\alias{shinyfactoranalysis}
\title{Factor Analysis using Dynamic Plots with Shiny}
\usage{
shinyfactoranalysis()
}
\value{
A shiny web app
}
\description{
This function calls a shiny app which performs factor analysis on a data set. Loadings and variance is calculated and model is tested for adequacy. Plots made to show
rotations of loadings.
}
\examples{
\dontrun{shinyMVNorm()}
}
