% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyMVNorm.R
\name{shinynormality}
\alias{shinynormality}
\title{Normality checking on univariate and multivariate data}
\usage{
shinynormality()
}
\value{
A shiny web app
}
\description{
This function calls a shiny app that contains dynamic plots,tables, and outputs relating to normality tests for univariate and muiltivariate
data, including proportion test and bivariate test. Plots are created to further aid in checking for normality, with multiple Q-Q plots.
}
\examples{
\dontrun{shinyMVNorm()}
}
