% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyrotations.R
\name{shinyrotations}
\alias{shinyrotations}
\title{Dynamic plots using Shiny}
\usage{
shinyrotations()
}
\value{
Shiny app
}
\description{
The app contains various widgets based on the input .csv file given. The user can then choose the variables to plot
and be able to calculate the correlations based on which point they want to drop. Additionally, the user will be able to see which root for
which the covariance is 0 given axes rotations.
}
\examples{
\dontrun{shinyrotations()}
}
