% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t2chart.R
\name{t2chart}
\alias{t2chart}
\title{T-Square Quality Control Chart for Multivariate Data}
\usage{
t2chart(x)
}
\arguments{
\item{x}{Multivariate data frame with 2+ variables}
}
\value{
A quality control ggplot chart
}
\description{
This function takes in multivariate data and calculates Hotelling's T-square then plots it against the number of observations to obtain a quality control
chart. The limits are obtained from chi-square quantiles for 95% and 99% confidence level.
}
\examples{
df <- data.frame(x = rnorm(20), y = rnorm(20))
t2chart(df)
}
