% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbarqc.R
\name{xbarqc}
\alias{xbarqc}
\title{X-bar Control Chart for Univariate Data}
\usage{
xbarqc(x)
}
\arguments{
\item{x}{univariate data}
}
\value{
A ggplot object
}
\description{
Function takes in a univariate data set and creates a control chart with horizontal lines showing the lower and upper limits being 3 standard deviations of mean.
Points outside of the limts will be out of control.
}
\examples{
data <- c(2,3,4,5,1,3,2)
xbarqc(data)
}
