% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariatettest.R
\name{bivariatettest}
\alias{bivariatettest}
\title{Bivariate T-test and Corresponding Confidence Intervals}
\usage{
bivariatettest(data, mu0, alpha = 0.05)
}
\arguments{
\item{data}{A data set with at least 2 qualitative variables}

\item{mu0}{The proposed null hypothesis mean values}

\item{alpha}{The rejection region.}
}
\value{
The result of a bivariate t-test. If the hypothesis is in the ellipse, then accept the null. Additionally,
eigenvalues and vectors, plot of confidence ellipse, major minor axis and ratio.
}
\description{
The function takes in a bivariate data set and tests it against a null hypothesis. It calculates Hotelling's T-square statistic and tests against a 50% chi-square quantile.
Additionally, a confidence ellipse around the points is created for a given alpha level.
}
\examples{
\dontrun{bivariatettest(rad.df, mu0 = c(.562,.589), alpha = .05)}
}
